Changed Project Group ID - Moved temporary group and artifact id from chdh.login.login to permanent org.mojodojocasahouse.extra - Changed module name from "login" to "extra" as in Expense(Ex.) Tracking(Tra.) Software
