+++ pom.xml
@@ -8,10 +8,10 @@
 		<version>3.1.3</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>mdch.login</groupId>
-	<artifactId>login</artifactId>
+	<groupId>org.mojodojocasahouse</groupId>
+	<artifactId>extra</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>login</name>
+	<name>ExTra</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
 		<java.version>20</java.version>
@@ -99,8 +99,8 @@
 				<configuration>
 					<formats>XML,CSV</formats>
 					<excludes>
-						<exclude>mdch/login/login/LoginApplication.class</exclude>
-						<exclude>mdch/login/login/dto/*</exclude>
+						<exclude>org/mojodojocasahouse/extra/ExtraApplication.class</exclude>
+						<exclude>org/mojodojocasahouse/extra/dto/*</exclude>
 					</excludes>
 				</configuration>
 				<executions>
