--- src/main/java/mdch/login/login/model/UsuarioEntity.java
@@ -1,4 +1,0 @@
-package mdch.login.login.model;
-
-public interface UsuarioEntity {
-}
