--- src/main/java/mdch/login/login/service/UsuarioService.java
@@ -1,12 +1,0 @@
-package mdch.login.login.service;
-
-import jakarta.validation.Valid;
-import mdch.login.login.dto.UsuarioRegistroDto;
-import mdch.login.login.dto.UsuarioRegistroResponseDto;
-import mdch.login.login.model.impl.Usuario;
-
-
-public interface UsuarioService {
-
-    UsuarioRegistroResponseDto registrarUsuario(UsuarioRegistroDto registroDTO);
-}
