+++ src/main/java/org/mojodojocasahouse/extra/ExtraApplication.java
@@ -1,13 +1,13 @@
-package mdch.login.login;
+package org.mojodojocasahouse.extra;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class LoginApplication {
+public class ExtraApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(LoginApplication.class, args);
+		SpringApplication.run(ExtraApplication.class, args);
 	}
 
 }
