+++ src/main/java/org/mojodojocasahouse/extra/dto/UsuarioRegistroResponseDto.java
@@ -1,4 +1,4 @@
-package mdch.login.login.dto;
+package org.mojodojocasahouse.extra.dto;
 
 import lombok.Data;
