+++ src/main/java/org/mojodojocasahouse/extra/model/UsuarioEntity.java
@@ -1,0 +1,4 @@
+package org.mojodojocasahouse.extra.model;
+
+public interface UsuarioEntity {
+}
